@@ -24,8 +24,8 @@ SOFTWARE.
 
 from __future__ import annotations
 
-from collections.abc import Iterable, Iterator
-from typing import TYPE_CHECKING, Callable, Generic, SupportsIndex, Union, cast, overload
+from collections.abc import Callable, Iterable, Iterator
+from typing import TYPE_CHECKING, Generic, SupportsIndex, cast, overload
 
 from typing_extensions import Self, TypeVar
 
@@ -67,7 +67,7 @@ class TransformerListProxy(Generic[T, K_co, V_co], list[T]):
         data = super().__getitem__(index)
         if isinstance(data, dict):
             # Narrow the type of data to Dict[str, Any]
-            raw_data: dict[K_co, V_co] = data
+            raw_data = cast(dict[K_co, V_co], data)
             result = self._transform_data(raw_data)
             super().__setitem__(index, result)
         else:
@@ -78,7 +78,7 @@ class TransformerListProxy(Generic[T, K_co, V_co], list[T]):
     def _transform_all(self):
         for index, entry in enumerate(self):
             if isinstance(entry, dict):
-                raw_data: dict[K_co, V_co] = entry
+                raw_data = cast(dict[K_co, V_co], entry)
                 result = self._transform_data(raw_data)
                 super().__setitem__(index, result)
 
@@ -94,7 +94,7 @@ class TransformerListProxy(Generic[T, K_co, V_co], list[T]):
     @overload
     def __getitem__(self, index: slice) -> list[T]: ...
 
-    def __getitem__(self, index: Union[SupportsIndex, slice]) -> Union[list[T], T]:
+    def __getitem__(self, index: SupportsIndex | slice) -> list[T] | T:
         if isinstance(index, slice):
             # This is a slice, so we need to handle each item in the slice and set it to the transformed data.
             # For each index in the slice, transform the data at that index then update the item list
